[Generator] move test to correct level
