+++ tests/Generator/Source/SomeFunction.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace ApiGen\Tests\Generator\TemplateGenerators\Source;
+namespace ApiGen\Tests\Generator\Source;
 
 function someFunction(): void
 {
