+++ tests/Generator/Source/SomeTrait.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace ApiGen\Tests\Generator\Source;
+
+trait SomeTrait
+{
+
+}
