--- tests/Generator/TemplateGenerators/Source/SomeTrait.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace ApiGen\Tests\Generator\TemplateGenerators\Source;
-
-trait SomeTrait
-{
-
-}
